requirements.txt
# Flask Backend
flask==3.0.0
flask-cors==4.0.0

# Database
flask-sqlalchemy==3.1.1

# AI - Gemini & Groq
google-generativeai==0.8.3
groq>=0.4.0

# TTS - Google
gTTS==2.5.3

# Video Generation
moviepy==1.0.3
Pillow==10.4.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment
python-dotenv==1.0.0

# Production server
gunicorn>=21.2.0
